pyproject.toml
[project]
name = "starlark-mcp-evals"
version = "0.1.0"
description = "Inspect AI evals for the Starlark MCP server"
requires-python = ">=3.11"
dependencies = [
    "inspect-ai>=0.3.185",
    "anthropic",
    "openai",
    "mcp",
    "pyyaml",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]

